mod_smtp_client: Use size_t for max message size to prevent overflow.
Overflow could occur if a mail server advertises support for very large
email messages, which previously could cause overflow, leading to false
positive send aborts, e.g.:

Message too large (2762 bytes, maximum is -2147483648)

Consistently use size_t instead of int to avoid this.

+++ include/mod_smtp_client.h
@@ -39,7 +39,7 @@ struct bbs_smtp_client {
 	const char *helohost;	/*!< Hostname to use for HELO/EHLO */
 	const char *hostname;	/*!< Hostname of remote SMTP server */
 	int caps;				/*!< Capabilities supported by remote SMTP server */
-	int maxsendsize;		/*!< Maximum size of messages accepted by remote SMTP server */
+	size_t maxsendsize;		/*!< Maximum size of messages accepted by remote SMTP server */
 	unsigned int secure:1;	/*!< Connection is secure */
 };
